Limit Switch overhaul
This changes the previously DigitalInput limit switches into values checks based on the voltage (that will indicate the servo's angle) to see if a "limit" has been reached. This mean that, while not a true limit switch, there are still active limits that can be checked. All limit switch variables are being kept in RobotMap for consistency. The limit switch check is being completed just before the main if/else statement concerning the number of blocks seen by the pixy.